still working on karyotype jobs.  Fixed major issue with loader that caused only half of the file to be loaded and incorrect numbering of the segments.

+++ hbase-genomes/src/main/java/org/lcsb/lu/igcsa/hbase/rows/ChromosomeRow.java
@@ -41,7 +41,7 @@ public class ChromosomeRow extends Row
     this.genome = name;
     }
 
-  public void addChromosomeInfo(String chrName, int length, int numSegments)
+  public void addChromosomeInfo(String chrName, long length, long numSegments)
     {
     super.addColumn( new Column("chr", "name", chrName) );
     super.addColumn( new Column("chr", "length", length) );
